Lifecycle.Massline (BigBrain v0.1.0)

The Massline User/Party context.

Functions

add_member(attrs \\ %{})

Adding user to a party

Input: %{"role" => "pleb",

    "party_id" => :binary_id,
    "user_id" => :}

Examples

iex> add_member(%{"role" => "pleb", "party_id" => "54635226-84df-4364-95f9-c5d3a3713baf", "user_id" => "39116133-3ded-455b-b972-198c54552cdf"}) {:ok, %Membership{}}

change_party(party, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking party changes.

Examples

iex> change_party(party) %Ecto.Changeset{data: %Party{}}

create_party(attrs \\ %{})

Creates a party.

Input: %{"founder_id" => :binary_id,

"name" => :string,
"banner" => :string}

Examples

iex> create_party(%{field: value}) {:ok, %Party{}}

iex> create_party(%{field: bad_value}) {:error, %Ecto.Changeset{}}

delete_party(party)

Deletes a party.

Examples

iex> delete_party(party) {:ok, %Party{}}

iex> delete_party(party) {:error, %Ecto.Changeset{}}

get_member!(attrs)

Return the membership between the a party and a user

Input: %{"party_id" => :binary_id,

      "user_id" => :binary_idz}

## Examples

iex> get_member!(%{"party_id" => "54635226-84df-4364-95f9-c5d3a3713baf", "user_id" => "39116133-3ded-455b-b972-198c54552cdf"}) %Membership{}

get_party!(id)

Gets a single party. Raises Ecto.NoResultsError if the Party does not exist.

Input: party_id: binary_id

Examples

iex> get_party!("7e66e29e-b302-4c66-915f-167dcf9c76e5") %Party{}

iex> get_party!("456") ** (Ecto.NoResultsError)

get_party_by_name!(name)

Gets party's id by passing in party's name Raises Ecto.NoResultsError if the Party does not exist.

Input: name: String

Examples

iex> get_party_by_name!("first party") %Party{}

iex> get_party_by_name!("456") ** (Ecto.NoResultsError)

get_user_by_name(name)

Gets user's id by passing in user's name

list_members(id)

Returns the list of members of a particular party

Input: party_id: :binary_id

Examples

iex> list_members(id) [%User{}, ...]

list_my_parties(id)

Returns the list of membership associated with the input user

Input: user_id: :binary_id

Examples

iex> list_my_parties(id) [%Membership{}, ...]

list_parties()

Returns the list of all parties

Input: None

Examples

iex> list_parties() [%Party{}, ...]

subtract_member(attrs \\ %{})

Remove user from a party

Input: %{"party_id" => :binary_id,

    "user_name" => :string}
    OR

Input: %{"party_id" => :binary_id,

    "user_id" => :binary_id}

Examples

iex> subtract_member(%{"party_id" => "54635226-84df-4364-95f9-c5d3a3713baf", "user_name" => "sky"})

update_party(party, attrs)

Updates a party.

Examples

iex> update_party(party, %{field: new_value}) {:ok, %Party{}}

iex> update_party(party, %{field: bad_value}) {:error, %Ecto.Changeset{}}