LifecycleWeb.Router.Helpers (BigBrain v0.1.0)

Module with named helpers generated from LifecycleWeb.Router.

Functions

admin_view_transition_path(conn_or_endpoint, action)

admin_view_transition_path(conn_or_endpoint, action, params)

admin_view_transition_url(conn_or_endpoint, action)

admin_view_transition_url(conn_or_endpoint, action, params)

billing_index_path(conn_or_endpoint, action)

billing_index_path(conn_or_endpoint, action, params)

billing_index_url(conn_or_endpoint, action)

billing_index_url(conn_or_endpoint, action, params)

echo_index_path(conn_or_endpoint, action)

echo_index_path(conn_or_endpoint, action, params)

echo_index_url(conn_or_endpoint, action)

echo_index_url(conn_or_endpoint, action, params)

echo_show_path(conn_or_endpoint, action, id)

echo_show_path(conn_or_endpoint, action, id, params)

echo_show_url(conn_or_endpoint, action, id)

echo_show_url(conn_or_endpoint, action, id, params)

journey_index_path(conn_or_endpoint, action)

journey_index_path(conn_or_endpoint, action, params)

journey_index_path(conn_or_endpoint, action, id, params)

journey_index_url(conn_or_endpoint, action)

journey_index_url(conn_or_endpoint, action, params)

journey_index_url(conn_or_endpoint, action, id, params)

journey_show_path(conn_or_endpoint, action, id)

journey_show_path(conn_or_endpoint, action, id, params)

journey_show_path(conn_or_endpoint, action, realm_name, journey_pointer, params)

journey_show_url(conn_or_endpoint, action, id)

journey_show_url(conn_or_endpoint, action, id, params)

journey_show_url(conn_or_endpoint, action, realm_name, journey_pointer, params)

live_dashboard_path(conn_or_endpoint, action)

live_dashboard_path(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action, page, params)

live_dashboard_path(conn_or_endpoint, action, node, page, params)

live_dashboard_url(conn_or_endpoint, action)

live_dashboard_url(conn_or_endpoint, action, params)

live_dashboard_url(conn_or_endpoint, action, page, params)

live_dashboard_url(conn_or_endpoint, action, node, page, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

party_index_path(conn_or_endpoint, action)

party_index_path(conn_or_endpoint, action, params)

party_index_path(conn_or_endpoint, action, party_name, params)

party_index_url(conn_or_endpoint, action)

party_index_url(conn_or_endpoint, action, params)

party_index_url(conn_or_endpoint, action, party_name, params)

party_show_path(conn_or_endpoint, action, party_name)

party_show_path(conn_or_endpoint, action, party_name, params)

party_show_url(conn_or_endpoint, action, party_name)

party_show_url(conn_or_endpoint, action, party_name, params)

path(data, path)

Generates the path information including any necessary prefix.

phase_index_path(conn_or_endpoint, action)

phase_index_path(conn_or_endpoint, action, params)

phase_index_path(conn_or_endpoint, action, phase_id, params)

phase_index_url(conn_or_endpoint, action)

phase_index_url(conn_or_endpoint, action, params)

phase_index_url(conn_or_endpoint, action, phase_id, params)

phase_show_path(conn_or_endpoint, action, phase_id)

phase_show_path(conn_or_endpoint, action, phase_id, params)

phase_show_path(conn_or_endpoint, action, phase_id, transition_id, params)

phase_show_url(conn_or_endpoint, action, phase_id)

phase_show_url(conn_or_endpoint, action, phase_id, params)

phase_show_url(conn_or_endpoint, action, phase_id, transition_id, params)

pow_registration_path(conn_or_endpoint, action)

pow_registration_path(conn_or_endpoint, action, params)

pow_registration_url(conn_or_endpoint, action)

pow_registration_url(conn_or_endpoint, action, params)

pow_session_path(conn_or_endpoint, action)

pow_session_path(conn_or_endpoint, action, params)

pow_session_url(conn_or_endpoint, action)

pow_session_url(conn_or_endpoint, action, params)

realm_index_path(conn_or_endpoint, action, id)

realm_index_path(conn_or_endpoint, action, id, params)

realm_index_url(conn_or_endpoint, action, id)

realm_index_url(conn_or_endpoint, action, id, params)

realm_show_path(conn_or_endpoint, action, party_name, id)

realm_show_path(conn_or_endpoint, action, party_name, id, params)

realm_show_url(conn_or_endpoint, action, party_name, id)

realm_show_url(conn_or_endpoint, action, party_name, id, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

trans_index_path(conn_or_endpoint, action)

trans_index_path(conn_or_endpoint, action, params)

trans_index_url(conn_or_endpoint, action)

trans_index_url(conn_or_endpoint, action, params)

transition_index_path(conn_or_endpoint, action)

transition_index_path(conn_or_endpoint, action, params)

transition_index_url(conn_or_endpoint, action)

transition_index_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.